Change normalization of the analytic solutions
Depending on whether one reads for instance a factor of 1/2 as part of
the noise operator or as part of the fluctuation amplitudes, one gets
different results for the filter function. We now settle on agreeing
with Cywinski et al. so that in the common dephasing Hamiltonian H =
\beta(t)\sigma_z/2, the noise operator is B = \sigma_z/2. This only
leaves a slight discrepancy of a factor of d in the definition of the
coherence function