Implement non-traceless bases
When dealing with different subspaces, one might want to generate an
operator basis that completely separates into elements that live
exclusively on the one and elements that live exclusively on the other
subspace. This is not possible when forcing all operator bases to have
an identity element (and thus the remaining elements traceless). The
core feature affected by this change is the way a complete basis is
generated from partial orthogonal elements. A switch now controls
whether a traceless or non-traceless is generated. Additionally, the
checks at the end of _full_from_partial() have been removed. The n
elements from which the basis is constructed now are the first n of
the final basis.

Including non-traceless bases also shone light on a bug in
calculate_error_transfer_matrix for a single qubit and the Pauli basis,
which did not include contributions from identity elements in case of a
cross-correlated spectrum. This induces non-unitality.