Implement bases whose elements have finite trace (#11)
* Implement non-traceless bases

When dealing with different subspaces, one might want to generate an
operator basis that completely separates into elements that live
exclusively on the one and elements that live exclusively on the other
subspace. This is not possible when forcing all operator bases to have
an identity element (and thus the remaining elements traceless). The
core feature affected by this change is the way a complete basis is
generated from partial orthogonal elements. A switch now controls
whether a traceless or non-traceless is generated. Additionally, the
checks at the end of _full_from_partial() have been removed. The n
elements from which the basis is constructed now are the first n of
the final basis.

Including non-traceless bases also shone light on a bug in
calculate_error_transfer_matrix for a single qubit and the Pauli basis,
which did not include contributions from identity elements in case of a
cross-correlated spectrum. This induces non-unitality.

* Cleanup and comments

* Add more tests

* Drop imaginary part of integrands when purely real

* Fix bug for U.ndim > 2

* Add comments

* Add utf-8 coding hint

* Increase atol to prevent test from sometimes failing

* Get n_oper_identifiers from pulse if available

* Check same dims of control and noise Hamiltonian

* Add function to more quickly compute exp(1j*x)

This is a major time sink in the calculation of the control matrix using
calculate_control_matrix_from_scratch(). Computing real and imaginary
part separately and writing to a preallocated array should make the
whole calculation faster by a factor of two.

* Update notebooks due to deprecation warnings

* Make QFT example work again

* Update RB example to work with the current version

* Add show_progressbar switch to error_transfer_matrix()

* Remove test that didn't make sense

* Project out subspace identity element in CNOT infid

* Simplify oper_equiv

* Replace :meth: with more accurate :func: sphinx refs

* Catch non-traceless bases in infidelity()

In this case, additional terms need to be considered.

* Set correct cutoff frequencies and fix scaling in test_infidelity_cnot()

* Sort imports