Make integral faster using np.expm1 and out kwarg
Using np.expm1 = np.exp - 1 is not only faster than calling cexp - 1 but
also more precise for small arguments, thus alleviating the need for
masking small values and manually setting the small-arg limit.

We can also precompute i*EdE by manually assigning the imaginary part of
a pre-allocated array using the out kwarg since we need it both in the
exponent and the denominator of the integral.