Make code more readable by allowing up to 99 characters
Since we name our variables and functions by the names of the
quantities they correspond to, lines just tend to get longer. Forcing
line breaks after 79 characters makes the code a mess.

The style implemented now tries to do as much as possible with 79
characters, but is more lenient when the code gets unreadable.

Also limit docstring width to 72 characters (most of the times) in
accordance with PEP8