Remove memory_parsimonious option for frequency shifts
Doesn't make much sense since we cannot factor the second order filter
function into control matrices as in the first order case.

In princinple, one can still loop the integration over one basis axis,
but this is a bit tedious and not pretty to implement, so we leave it
for now.